Fix potential race condition when consuming OkHttp response body.
